+++ Frameworks/AIUtilities Framework/Source/AIFontManagerAdditions.h
@@ -16,6 +16,6 @@
 @interface NSFontManager (AIFontManagerAdditions)
 
 //find a font by family name, using case-insensitive matching.
-- (NSFont *)fontWithFamilyInsensitively:(NSString *)name traits:(NSFontTraitMask)fontTraitMask weight:(int)weight size:(float)size;
+- (NSFont *)fontWithFamilyInsensitively:(NSString *)name traits:(NSFontTraitMask)fontTraitMask weight:(NSInteger)weight size:(CGFloat)size;
 
 @end
