+++ Frameworks/AIUtilities Framework/Source/AIImageDrawingAdditions.h
@@ -17,15 +17,15 @@ typedef enum {
 
 - (void)tileInRect:(NSRect)rect;
 - (NSImage *)imageByScalingToSize:(NSSize)size;
-- (NSImage *)imageByFadingToFraction:(float)delta;
-- (NSImage *)imageByScalingToSize:(NSSize)size fraction:(float)delta;
+- (NSImage *)imageByFadingToFraction:(CGFloat)delta;
+- (NSImage *)imageByScalingToSize:(NSSize)size fraction:(CGFloat)delta;
 - (NSImage *)imageByScalingForMenuItem;
-- (NSImage *)imageByScalingToSize:(NSSize)size fraction:(float)delta flipImage:(BOOL)flipImage proportionally:(BOOL)proportionally allowAnimation:(BOOL)allowAnimation;
+- (NSImage *)imageByScalingToSize:(NSSize)size fraction:(CGFloat)delta flipImage:(BOOL)flipImage proportionally:(BOOL)proportionally allowAnimation:(BOOL)allowAnimation;
 //+ (NSImage *)imageFromGWorld:(GWorldPtr)gWorldPtr;
-- (NSRect)drawRoundedInRect:(NSRect)rect radius:(float)radius;
-- (NSRect)drawRoundedInRect:(NSRect)rect fraction:(float)fraction radius:(float)radius;
-- (NSRect)drawRoundedInRect:(NSRect)rect atSize:(NSSize)size position:(IMAGE_POSITION)position fraction:(float)fraction radius:(float)radius;
-- (NSRect)drawInRect:(NSRect)rect atSize:(NSSize)size position:(IMAGE_POSITION)position fraction:(float)fraction;
+- (NSRect)drawRoundedInRect:(NSRect)rect radius:(CGFloat)radius;
+- (NSRect)drawRoundedInRect:(NSRect)rect fraction:(CGFloat)fraction radius:(CGFloat)radius;
+- (NSRect)drawRoundedInRect:(NSRect)rect atSize:(NSSize)size position:(IMAGE_POSITION)position fraction:(CGFloat)fraction radius:(CGFloat)radius;
+- (NSRect)drawInRect:(NSRect)rect atSize:(NSSize)size position:(IMAGE_POSITION)position fraction:(CGFloat)fraction;
 - (NSRect)rectForDrawingInRect:(NSRect)rect atSize:(NSSize)size position:(IMAGE_POSITION)position;
 
 @end
